requirements.txt
streamlit
streamlit-tags
pandas
selenium
html2text
beautifulsoup4
python-dotenv
pydantic
google-generativeai
